refactor: share rpc endpoint transport
